End-to-End-IoT-Sensor-Data-Analytics(Real-Time)/requirements.txt
kafka-python==2.0.2
pyspark==3.5.4
